# JIT: optimize term to int using types