Continuation of:
-  #1770 
- #1833

First optimization using type information: simplify some term_to_int implementations.

Current benchmark:
|test                 |JIT disabled (c06edb2bca8dacc8c17661575b39bcfc2c5922ea)|HEAD~1, JIT enabled (9b14b166899f3248189b71c6c157c6ae110df6c4)|This PR (b775240224693b873ed69c74f08673e0d30f9ec3) |
|--------------|---------------- |------------------------|--------|
|erlang-tests   | 11.45s                  | 9.95s  | 9.94s |
|benchmark (entire run)                      | 17.03s                 | 15.35s | 15.52s |
|benchmark : pingpong_speed_test | 3.59s                 | 3.23s | 3.38s |
|benchmark : prime_speed_test       |  0.26s                            | 0.17s | 0.17s |
|benchmark : sudoku_solution_test |  0.26s                            | 1.34s | 1.36s |
|benchmark : sudoku_puzzle_test   |    10.51s                            | 9.18s | 9.22s |

These changes are made under both the "Apache 2.0" and the "GNU Lesser General
Public License 2.1 or later" license terms (dual license).

SPDX-License-Identifier: Apache-2.0 OR LGPL-2.1-or-later